# api/openapi.yaml
openapi: 3.0.3
info:
  title: JellyWatch API
  description: API for managing media library cleanup operations
  version: 2.0.0
  license:
    name: GPL-3.0
servers:
  - url: /api/v1
    description: API v1
paths:
...
  /webhooks/jellyfin:
    post:
      operationId: jellyfinWebhook
      summary: Receive Jellyfin webhook events
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Event accepted
        '400':
          description: Invalid payload
        '401':
          description: Invalid or missing webhook secret
...